refactoring to remove analytic expansion into its own object
