+++ nav2_smac_planner/CMakeLists.txt
@@ -69,6 +69,7 @@ set(dependencies
 add_library(${library_name} SHARED
   src/smac_planner_hybrid.cpp
   src/a_star.cpp
+  src/analytic_expansion.cpp
   src/node_hybrid.cpp
   src/node_lattice.cpp
   src/costmap_downsampler.cpp
@@ -86,6 +87,7 @@ ament_target_dependencies(${library_name}
 add_library(${library_name}_2d SHARED
   src/smac_planner_2d.cpp
   src/a_star.cpp
+  src/analytic_expansion.cpp
   src/node_hybrid.cpp
   src/node_lattice.cpp
   src/costmap_downsampler.cpp
@@ -103,6 +105,7 @@ ament_target_dependencies(${library_name}_2d
 add_library(${library_name}_lattice SHARED
   src/smac_planner_lattice.cpp
   src/a_star.cpp
+  src/analytic_expansion.cpp
   src/node_hybrid.cpp
   src/node_lattice.cpp
   src/costmap_downsampler.cpp
